+++ tool/src/ase.ts
@@ -21,6 +21,7 @@ import MetaCommand                 from "./ase-meta.js"
 import CompatCommand               from "./ase-compat.js"
 import DiagramCommand              from "./ase-diagram.js"
 import WorktreeCommand             from "./ase-worktree.js"
+import MintCommand                 from "./ase-mint.js"
 import pkg                         from "../package.json" with { type: "json" }
 
 /*  type of top-level (global) options  */
@@ -73,6 +74,7 @@ const main = async (): Promise<void> => {
     new CompatCommand().register(program)
     new DiagramCommand(log).register(program)
     new WorktreeCommand().register(program)
+    new MintCommand().register(program)
 
     /*  parse program arguments  */
     await program.parseAsync(process.argv)